% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_grafi.R
\name{cont_grafi}
\alias{cont_grafi}
\title{cont_grafi}
\usage{
cont_grafi(df, coluna)
}
\arguments{
\item{coluna}{}
}
\description{
cont_grafi
}
\examples{
cont_grafi(mtcars, 'am')
cont_grafi(mtcars, 'gear')
# ggsave("nome_grafico.png", height=15, width=20, units="cm", dpi= 600)

}
