% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_diagnostic.R
\name{glm_diagnostic}
\alias{glm_diagnostic}
\title{glm_diagnostic}
\usage{
glm_diagnostic(modelo)
}
\arguments{
\item{modelo}{}
}
\description{
grafico de analise de residuos para modelos generalizados
}
\examples{
modelo = glm(am ~ vs, data = mtcars, family = binomial())
glm_diagnostic(modelo)
}
