% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnostic.R
\name{lm_diagnostic}
\alias{lm_diagnostic}
\title{lm_diagnostic}
\usage{
lm_diagnostic(modelo)
}
\arguments{
\item{modelo}{}
}
\description{
grafico de analise de residuos para modelos lineares
}
\examples{
modelo_lm = lm(mpg ~ vs, data = mtcars, family = binomial())
lm_diagnostic(modelo_lm)
# modelo_lmer <- lmerTest::lmer(response ~ fixed_effects + (1|group), data = dff)
# lm_diagnostic(modelo_lmer)
}
