% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_normalidade.R
\name{teste_normalidade}
\alias{teste_normalidade}
\title{teste_normalidade}
\usage{
teste_normalidade(df, coluna, qtd_bins = 20, cor_esc = 1, plot_qqplot = T)
}
\arguments{
\item{plot_qqplot}{}
}
\description{
grafico de normalidade mais teste de hipotese
}
\examples{
teste_normalidade(mtcars, 'mpg', cor_esc = 2)
}
